Changed HDF5 Behavior in data.py
I wanted the table output in the data.h5 file to be a bit more structured.
Looked up how to create a hdf5 table from an dictionary rather than a
predefined class. In this way, I could have a table with each row representing
the concentration of one of the components.