Added refs_name function
I added a new function to the gcms.py file that extracts the reference list
file. This was being used in both data.py and calibration.py, so consolidating
it seemed like the best option. Now it will be easier to use in IPython as
well.

In addition, it seemed that my last commit caused some problems. Renamed the
data table with a better Python name, but also added a longer description to
the table so people will know it is concentration data.