Added extra data to data file
I wanted to export the raw integration data for each calibration region to the
data HDF5 file. In addition, I also added the percentage, so that I could see
the percent composition from each species. This might be useful at some later
date, especially if a lot of extra components are being fit to a peak that I
know is only one species.