Add background time selection for ref build
Modified the reference file builder with a toggle switch for adding a
background spectrum for fitting purposes. In addition, I also added another
keyword argument so that you could set the time that you'd like to use for
background subtraction. This time is then converted to an index from the
AIAFile.time array. The AIAFile class saves the index, so that you can check
it later, if necessary.