Added command args for bkg to cal script
The calibration automation script was modified to accept command line
arguements. The first argument '--nobkg' turns off the useage of a MS slice as
a background for fitting. The second argument, '--bkg_time', sets the time
for the background MS slice. The values of these two arguments is saved into
the hdf5 calibration file as attributes of the 'cals' table. That way they can
be queried at a later time.