Modified data.py for command line args
Added the command line argument stuff to data.py as well. This impacts the
reference building command for the AIA file processing, which was outlined in
a previous commit.

I noticed that the command line arguments were going to be the same as for
calibration.py, so I added the argument parser stuff to gcms.py. Now there
will be a unified location for these commands.