Modified argument parser for calibration.py
In the last commit, I moved the argument parsing stuff to gcms.py, so that it
would be more global for other scripts. I modified calibration.py so that it
uses the argument parser from gcms.py rather than a local copy of the code.

Otherwise, everything else works the same.