Fixed some problems with data.py HDF5 file
The script data.py was generating a HDF5 output file that contained too many
columns of compounds. It was making columns for reference spectra files, even
if they did not have corresponding calibration information. I modified a
couple sections of this script to generate the columns from the calibration
data file instead, so that these files would contain info for the same
compounds.