Modify ref_build to take ref file not list
I modified the ref_build function of that AIAFile class, so that you can pass
in different types of reference files rather than a list of file names. This
was to accomodate the reference files from AMDIS, which put all reference info
into a single file. This required some modifications to the data.py and
calibration.py files as well.