Tried some memory/speed optimization with data types
The intensity array was getting forced to a double precision floating point
array. However, the input array was already float32. Just changed it so that
there was no type conversion. This did not help with speed, but will most
likely help with memory usage.

For the groupby operation, I changed the masses to be integers instead of
floats. That made a small difference in execution speed. Something to think
about when I optimize that section of the code later.