Integration fix, reference comments, and unicode
This commit has several fixes rolled into one.

First of all, While working with some GCMS data, I noticed that the
integrations I was calculating were not exactly correct. I was using the
least squares coefficients directly as the amount of a particular compound.
However, I should have been multiplying these coefficients by the reference MS
and then summing the total MS intensity. This is more in line with what is
already done to get the TIC plot.

I also made it possible to have comments in a reference MSL file, so that I
could comment out selected reference compounds.

Finally, I added a __future__ import for unicode literals. There are some
unusual characters in the MSL file, and I'll need to work on this encoding
business to figure that out.