Refactored calibration.py
I refactored calibration.py in order to make it easier to modify in the
future. More things are factored into functions.

I also made it so that there is only one csv file with reference information.
This removes a lot of the files that were laying around in the directory.