Fixed serious error in reference array
I made a serious mistake while generating the reference array for my GCMS
fitting. If masses for the reference spectra were below the lowest mass that
was being detected, they were wrapped around to the other end of the mass
spectrum, which meant that I had intensities at very high masses.

Added a couple of lines of code to fix that issue, and things seem to be
better now.