Fixed multiproc and cleanup
I fixed the multiprocessing stuff so that if the number of processors is 1, it
will not use multiprocessing.Pool, which doesn't work in the interactive
session. Also added the number of processors as a command line argument.

Cleaned up a bunch of stuff having to do with integrations. Don't need to
return anything from the integration function.