Replace Pandas Groupby with itertools
There was a small bit of code in the AIA file processing function that grouped
the intensity values by masses and then averaged the intensity values. I did
this initially with Pandas because it was the easiest solution. I replaced
that code with the groupby function from itertools, which is in the standard
library. This required a custom key extraction function for my AIAFile object.
This is much faster than the Pandas version, with a speedup of around 8x.
Also, dropped the Pandas import since that is unnecessary now.