Added a retention time filter to nnls function
I needed a way to filter potential peak fitting to a selection of compounds
based on their known experimental retention times. I added some code to the
nnls function in my AIAFile object to do just that. It requires that the
reference file has retention time information included with the reference
spectra.

Seems to work okay on several tests.