Add dynamic class generator
I added a function that dynamically generates a new class that has all of the
functions needed to process the GCMS data set.

This required a substantial reorganization of my code base to accomodate this
type of constructor. I also had to add a __reduce__ function to make pickling
possible.